% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Chem.Mean.YearDF}
\alias{Chem.Mean.YearDF}
\title{Physical Chemical Data for Bavarian Lakes}
\format{
A data frame with 53940 rows and 4 variables:
\describe{
  \item{Var1}{Lake Name}
  \item{Var2}{Year of measurements}
  \item{Var3}{Chemical physical variables}
  \item{value}{values}
  ...
}
}
\source{
\url{http://www.lfu.bayern.de/}
}
\usage{
Chem.Mean.YearDF
}
\description{
A dataset containing annual means of physical chemical measurements
for Bavarian lakes.
Data were prepaired in "DATASET.R"
}
\keyword{datasets}
