% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearmodel_pvalue.R
\name{lmp}
\alias{lmp}
\title{Returns p-value of a linear model}
\usage{
lmp(modelobject)
}
\arguments{
\item{modelobject}{}
}
\value{

}
\description{
Extracts p-value out of a linear model produced by function lm(); function found somewhere in the Internet
}
\details{
Title
}
\examples{
bla
}
